refactored into L.lua file, new TOD billing features
Moved most of the lua code out of the I.xml file into functions in the L.lua file.  Added new actions, StartPeak and EndPeak, which can be useful if you have Time of Day billing.  Call these appropriately and the plugin will split the usage into KWHPeak and KWHOffPeak so you know how much you will be billed.  Also added KWHDeliveredPerPeriod so you can monitor the delivered total, not just net total.  Some utilities charge for delivery regardless of the net, so this is a handy number to monitor.  All of these are reset when the ResetKWH action is called.  Ideally this should be called at the end of your billing period.