fixed bug in resetKWH, updates to CommFailure
resetKWH() had getVar's with the service as a device (wrong!).  Also updated the CommFailure logging to try to catch the timestamp of the first failure (in case there is a repeating problem you want to know when the first failure occurred), and also to only say communication is good when the parsing finds good data, not just that a valid response was received.