luup.set_failure and HAN_HWADDR
Don't use luup.set_failure with UI7 - user reports of issues, and I have no way to test.
Use a device variable for the HWADDR and only pick it up once.  Eagle 200 has nasty habit of occasionally returning an empty string on the device details request - if this coincides with a Vera luup restart, the plugin is stuck.  Instead, get it the first time, and keep it in a device variable, and don't ask again.  If the Eagle fails the very first time, the user can restart luup and to resolve it.